Rename writer/ to io/ and move ccio.py there
