+++ src/cclib/io/__init__.py
@@ -1,7 +1,7 @@
 # This file is part of cclib (http://cclib.github.io), a library for parsing
 # and interpreting the results of computational chemistry packages.
 #
-# Copyright (C) 2014, the cclib development team
+# Copyright (C) 2014-2016, the cclib development team
 #
 # The library is free software, distributed under the terms of
 # the GNU Lesser General Public version 2.1 or later. You should have
@@ -10,11 +10,15 @@
 
 """Contains all writers for standard chemical representations"""
 
-
 from .cjsonwriter import CJSON
 from .cmlwriter import CML
 from .xyzwriter import XYZ
 
 # This allows users to type:
-#     from cclib.writer import ccwrite
-from .ccwrite import ccwrite
+#   from cclib.io import ccopen
+#   from cclib.io import ccread
+#   from cclib.io import ccwrite
+from .ccio import ccopen
+from .ccio import ccread
+from .ccio import ccwrite
+
