+++ src/cclib/io/cjsonreader.py
