+++ src/cclib/io/cjsonwriter.py
