+++ src/cclib/io/filewriter.py
