+++ src/cclib/parser/__init__.py
@@ -3,7 +3,7 @@
 # This file is part of cclib (http://cclib.github.io), a library for parsing
 # and interpreting the results of computational chemistry packages.
 #
-# Copyright (C) 2006-2014, the cclib development team
+# Copyright (C) 2006-2016, the cclib development team
 #
 # The library is free software, distributed under the terms of
 # the GNU Lesser General Public version 2.1 or later. You should have
@@ -30,11 +30,5 @@ from .nwchemparser import NWChem
 from .orcaparser import ORCA
 from .psiparser import Psi
 from .qchemparser import QChem
-
-# This allow users to type:
-#   from cclib.parser import ccopen
-#   from cclib.parser import ccread
-from .ccio import ccopen
-from .ccio import ccread
 
 from .data import ccData
