+++ src/scripts/ccwrite
@@ -17,8 +17,9 @@ import logging
 import os.path
 import sys
 
-from cclib.parser import ccData, ccopen
-from cclib.writer import ccwrite
+from cclib.parser import ccData
+from cclib.io import ccopen
+from cclib.io import ccwrite
 
 
 def main():
