+++ src/scripts/cda
@@ -3,7 +3,7 @@
 # This file is part of cclib (http://cclib.github.io), a library for parsing
 # and interpreting the results of computational chemistry packages.
 #
-# Copyright (C) 2007-2014, the cclib development team
+# Copyright (C) 2007-2016, the cclib development team
 #
 # The library is free software, distributed under the terms of
 # the GNU Lesser General Public version 2.1 or later. You should have
@@ -19,7 +19,7 @@ import logging
 
 import numpy
 
-from cclib.parser import ccopen
+from cclib.io import ccopen
 from cclib.method import CDA
